#27 refactor: rename variables
