+++ fastapi/src/services/service_factory.py
@@ -1,6 +1,6 @@
 from fastapi import Depends
 
-from db.elastic import ElasticDB, get_elastic
+from db.elastic_dao import ElasticDAO, get_elastic
 from services.base_service import BaseService
 
 
@@ -8,7 +8,7 @@ def service_for(service_name: str):
     """Возвращает FastAPI-зависимость, создающую нужный сервис."""
 
     def service_provider(
-        elastic: ElasticDB = Depends(get_elastic)
+        elastic: ElasticDAO = Depends(get_elastic)
     ) -> BaseService:
         return BaseService.get_instance(service_name, elastic)
